+++ src/cli/action.rs
@@ -1,10 +1,8 @@
-use crate::core::messenger::msg;
-
-use super::RunInfo;
+use super::{Instance, Run};
 
 /// Action that can be done with cli
 pub enum Action {
     /// Sends the given message
-    Message(Vec<msg::Message>),
-    RunDetached(RunInfo),
+    Instance(Instance),
+    RunDetached(Run),
 }